requirements.txt
# Spider core dependencies
scrapy>=2.5.0
selenium>=4.0.0
beautifulsoup4>=4.9.0

# Web parsing
lxml>=4.6.0
html5lib>=1.1

# Utilities
requests>=2.25.0

# Optional: Better Chrome driver management
webdriver-manager>=3.8.0

# FastAPI and web server dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
jinja2>=3.0.0

# HTML to Markdown conversion tools
html2text>=2020.1.16
markdownify>=0.11.6

# PostgreSQL database support
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0

# Development tools (optional)
ipython>=7.0.0